Internal refactoring: we now use a more robust method to identify tests that are steps of the same test.
Now all methods used to get a unique test node id with or without removal of parameters are in steps_commons.py.
Both parametrize and generator methods now use the same way to identify the tests that are steps belonging to the same test.

This fixes some bugs that were happening on edge cases where several parameters had the same string id representation (or one was a substring of the other). Fixed #21

Added a corresponding test